Added solution of the third task

+++ app/controllers/dashboard/users_controller.rb
@@ -1,8 +1,13 @@
 # File for Dashboard::UsersController class
 class Dashboard::UsersController < Dashboard::BaseController
   def destroy
+    authorize current_user
     current_user.destroy
     redirect_to login_path, notice: 'Пользователь успешно удален.'
+  end
+
+  def admin_list
+    authorize current_user
   end
 
   private
